feat(auth): implement user authentication and management functions including login, registration, password update, and token handling
